flask/requirements.txt
Flask
Flask-Session
gspread
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
